removed all localhost to achieve relative url

+++ models/user.js
@@ -1,6 +1,6 @@
 module.exports = function(sequelize, DataTypes) {
   var model = sequelize.define("user", {
-    twitter_id: DataTypes.INTEGER
+    twitter_id: DataTypes.BIGINT
     // name: DataTypes.STRING
     // username: DataTypes.STRING,
     // password_digest: DataTypes.STRING
